pid test

+++ firmware/car/targets/MKV/algorithms/pid.hpp
@@ -19,7 +19,7 @@ class PID {
     public:
         PID(float kp = -1.8f, float ki = -0.0001f, float kd = 0.0f,
             float maxValue = 100.0f, float bias = 0.0f);
-        int32_t calculate(int32_t setpoint, int32_t current);
+        int32_t calculate(float setpoint, float current);
 
         void setKp(float kp);
         float getKp() const;
